adds dequeing customers who cancel after entering
the algorithm to delete from inside the queue is to either adjust
the head and tail pointers or to mark the value as deleted. then
on dequeue, the value is checked against the known mark and skipped
if it was meant to be deleted.

lots of logging added to debug the implementation of the queue.

customer sends a synchronous ajax request when the page unloads in
case they leave without pressing the cancel button

added error checking in multiple places on the server implementation